# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "MediaType": {
        "enum": [
          "audio",
          "video",
          "none"
        ],
        "title": "MediaType",
        "type": "string"
      },
      "MeetingCreate": {
        "additionalProperties": false,
        "properties": {
          "channel_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel Id"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "participants": {
            "items": {
              "maxLength": 200,
              "minLength": 1,
              "type": "string"
            },
            "title": "Participants",
            "type": "array"
          },
          "segments": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SegmentIn"
                },
                "minItems": 1,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Segments"
          },
          "source": {
            "$ref": "#/components/schemas/MeetingSource",
            "default": "manual"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "title": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "MeetingCreate",
        "type": "object"
      },
      "MeetingDetail": {
        "properties": {
          "action_item_counts": {
            "$ref": "#/components/schemas/ActionItemCountsRead"
          },
          "channel_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel Id"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "duration_ms": {
            "title": "Duration Ms",
            "type": "integer"
          },
          "has_media": {
            "title": "Has Media",
            "type": "boolean"
          },
          "host": {
            "$ref": "#/components/schemas/UserRef"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "title": "Keywords",
            "type": "array"
          },
          "media_type": {
            "$ref": "#/components/schemas/MediaType"
          },
          "participant_count": {
            "title": "Participant Count",
            "type": "integer"
          },
          "participants": {
            "items": {
              "$ref": "#/components/schemas/ParticipantRead"
            },
            "title": "Participants",
            "type": "array"
          },
          "source": {
            "$ref": "#/components/schemas/MeetingSource"
          },
          "speakers": {
            "items": {
              "$ref": "#/components/schemas/SpeakerRead"
            },
            "title": "Speakers",
            "type": "array"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/MeetingStatus"
          },
          "suggested_tags": {
            "items": {
              "type": "string"
            },
            "title": "Suggested Tags",
            "type": "array"
          },
          "summary_status": {
            "enum": [
              "none",
              "ready",
              "stale",
              "generating"
            ],
            "title": "Summary Status",
            "type": "string"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TagRead"
            },
            "title": "Tags",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "started_at",
          "duration_ms",
          "host",
          "participant_count",
          "action_item_counts",
          "keywords",
          "tags",
          "has_media",
          "description",
          "participants",
          "speakers",
          "channel_id",
          "source",
          "status",
          "media_type",
          "summary_status",
          "suggested_tags"
        ],
        "title": "MeetingDetail",
        "type": "object"
      },
      "MeetingListItem": {
        "description": "Light row for lists: never carries the transcript.",
        "properties": {
          "action_item_counts": {
            "$ref": "#/components/schemas/ActionItemCountsRead"
          },
          "duration_ms": {
            "title": "Duration Ms",
            "type": "integer"
          },
          "has_media": {
            "title": "Has Media",
            "type": "boolean"
          },
          "host": {
            "$ref": "#/components/schemas/UserRef"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "title": "Keywords",
            "type": "array"
          },
          "overview_preview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overview Preview"
          },
          "participant_count": {
            "title": "Participant Count",
            "type": "integer"
          },
          "participants": {
            "items": {
              "$ref": "#/components/schemas/ParticipantRef"
            },
            "maxItems": 5,
            "title": "Participants",
            "type": "array"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TagRead"
            },
            "title": "Tags",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "started_at",
          "duration_ms",
          "host",
          "participant_count",
          "action_item_counts",
          "keywords",
          "tags",
          "has_media",
          "participants",
          "overview_preview"
        ],
        "title": "MeetingListItem",
        "type": "object"
      },
      "MeetingSort": {
        "enum": [
          "-started_at",
          "started_at",
          "title",
          "-duration_ms"
        ],
        "title": "MeetingSort",
        "type": "string"
      },
      "MeetingSource": {
        "enum": [
          "seed",
          "upload",
          "paste",
          "manual",
          "capture",
          "calendar"
        ],
        "title": "MeetingSource",
        "type": "string"
      },
...
      "MeetingUpdate": {
        "additionalProperties": false,
        "description": "Partial update; unset fields are untouched, so null can clear description/channel.",
        "properties": {
          "channel_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel Id"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "participants": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ParticipantInput"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Participants"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 300,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "MeetingUpdate",
        "type": "object"
      },
...
      "Page_MeetingListItem_": {
        "properties": {
          "has_next": {
            "readOnly": true,
            "title": "Has Next",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/MeetingListItem"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "total_pages": {
            "readOnly": true,
            "title": "Total Pages",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total",
          "total_pages",
          "has_next"
        ],
        "title": "Page[MeetingListItem]",
        "type": "object"
      },
...
      "SummaryRead": {
        "properties": {
          "generated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At"
          },
          "is_stale": {
            "title": "Is Stale",
            "type": "boolean"
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "title": "Keywords",
            "type": "array"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "notes": {
            "items": {
              "$ref": "#/components/schemas/NoteGroupRead"
            },
            "title": "Notes",
            "type": "array"
          },
          "outline": {
            "items": {
              "$ref": "#/components/schemas/OutlineEntryRead"
            },
            "title": "Outline",
            "type": "array"
          },
          "overview": {
            "title": "Overview",
            "type": "string"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          }
        },
        "required": [
          "overview",
          "keywords",
          "outline",
          "notes",
          "provider",
          "model",
          "generated_at",
          "is_stale"
        ],
        "title": "SummaryRead",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Fireflies Clone API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/meetings": {
      "get": {
        "operationId": "list_meetings",
        "parameters": [
          {
            "description": "Prefix `-` means descending.",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/MeetingSort",
              "default": "-started_at",
              "description": "Prefix `-` means descending."
            }
          },
          {
            "description": "Matches title, overview, people, transcript.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Matches title, overview, people, transcript.",
              "title": "Q"
            }
          },
          {
            "description": "Participant name contains.",
            "in": "query",
            "name": "participant",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Participant name contains.",
              "title": "Participant"
            }
          },
          {
            "description": "Inclusive, UTC day.",
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive, UTC day.",
              "title": "Date From"
            }
          },
          {
            "description": "Inclusive, UTC day.",
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive, UTC day.",
              "title": "Date To"
            }
          },
          {
            "description": "Tag id; repeat for any-of.",
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "integer"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Tag id; repeat for any-of.",
              "title": "Tag"
            }
          },
          {
            "description": "Channel id.",
            "in": "query",
            "name": "channel",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Channel id.",
              "title": "Channel"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "all",
              "enum": [
                "all",
                "hosted",
                "shared",
                "uploads"
              ],
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "completed",
              "enum": [
                "completed",
                "upcoming"
              ],
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Clamped to 100.",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Clamped to 100.",
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_MeetingListItem_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input; `details.errors[].loc` is the field path."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A dependency is unavailable, e.g. the database is unseeded."
          }
        },
        "summary": "List meetings",
        "tags": [
          "meetings"
        ]
      },
      "post": {
        "operationId": "create_meeting",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeetingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input; `details.errors[].loc` is the field path."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A dependency is unavailable, e.g. the database is unseeded."
          }
        },
        "summary": "Create a meeting",
        "tags": [
          "meetings"
        ]
      }
    },
    "/api/v1/meetings/{meeting_id}": {
      "delete": {
        "operationId": "delete_meeting",
        "parameters": [
          {
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "title": "Meeting Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The resource does not exist."
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The meeting was soft-deleted; restore it to use it again."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input; `details.errors[].loc` is the field path."
          }
        },
        "summary": "Soft-delete a meeting",
        "tags": [
          "meetings"
        ]
      },
      "get": {
        "operationId": "get_meeting",
        "parameters": [
          {
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "title": "Meeting Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The resource does not exist."
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The meeting was soft-deleted; restore it to use it again."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input; `details.errors[].loc` is the field path."
          }
        },
        "summary": "Get a meeting",
        "tags": [
          "meetings"
        ]
      },
      "patch": {
        "operationId": "update_meeting",
        "parameters": [
          {
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "title": "Meeting Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeetingUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The resource does not exist."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The request conflicts with current state, e.g. a duplicate name."
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The meeting was soft-deleted; restore it to use it again."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input; `details.errors[].loc` is the field path."
          }
        },
        "summary": "Update a meeting",
        "tags": [
          "meetings"
        ]
      }
    },
...
    "/api/v1/meetings/{meeting_id}/restore": {
      "post": {
        "operationId": "restore_meeting",
        "parameters": [
          {
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "title": "Meeting Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The resource does not exist."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input; `details.errors[].loc` is the field path."
          }
        },
        "summary": "Restore a soft-deleted meeting",
        "tags": [
          "meetings"
        ]
      }
    },
    "/api/v1/meetings/{meeting_id}/summary": {
      "get": {
        "operationId": "get_summary",
        "parameters": [
          {
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "title": "Meeting Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummaryRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The resource does not exist."
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The meeting was soft-deleted; restore it to use it again."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input; `details.errors[].loc` is the field path."
          }
        },
        "summary": "Get a meeting's summary",
        "tags": [
          "summaries"
        ]
      }
    },
    "/api/v1/meetings/{meeting_id}/summary/regenerate": {
      "post": {
        "operationId": "regenerate_summary",
        "parameters": [
          {
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "title": "Meeting Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummaryRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The resource does not exist."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The request conflicts with current state, e.g. a duplicate name."
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The meeting was soft-deleted; restore it to use it again."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input; `details.errors[].loc` is the field path."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "AI rate limit exceeded; retry later."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The AI provider failed; `code` is AI_UNAVAILABLE."
          }
        },
        "summary": "Regenerate the summary with AI",
        "tags": [
          "summaries"
        ]
      }
    },
...
  }
}